Добавлены тесты PerfectlyBalancedTree.h

+++ Nosov_1k_2s_Lab_3/ConstantsBoostTest.h
@@ -11,4 +11,4 @@ int SizeConstArrayIntHeap = 11;
 float ConstArrayFloatHeap[11] = { 2.0, 5.0, 100.0, 1500.0, 5050.0, 13.0, 14.0, 27.0, 70.0, 150.0, 40.0 };
 int SizeConstArrayFloatHeap = 11;
 
-float Epsilon = 0.1;
+float Epsilon = (float)0.1;
